moved to new header and footer
